Moved _mata_builtins.py to pygments.
